requirements.txt
torch>=2.0.0
opencv-python
diffusers==0.26.3
accelerate
peft==0.7.1
transformers==4.37.2
safetensors
controlnet-aux==0.0.3
fastapi==0.108.0
uvicorn
python-multipart
python-socketio==5.7.2
typer
pydantic>=2.5.2
rich
loguru
yacs
piexif==1.1.3
omegaconf
easydict
gradio
typer-config==1.4.0
Pillow==9.5.0
ultralytics
flask
flask_cors